BJS2-54229 Добавление картинок к посту. Без тестов Fixed

+++ src/main/java/faang/school/postservice/service/s3/S3Service.java
@@ -1,10 +1,12 @@
 package faang.school.postservice.service.s3;
 
+import com.amazonaws.AmazonServiceException;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.amazonaws.services.s3.model.S3Object;
+import faang.school.postservice.exception.IntegrationException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -24,10 +26,8 @@ public class S3Service {
     private String bucketName;
 
     public String uploadFile(MultipartFile file, String folder) {
-
-        long fileSize = file.getSize();
         ObjectMetadata objectMetadata = new ObjectMetadata();
-        objectMetadata.setContentLength(fileSize);
+        objectMetadata.setContentLength(file.getSize());
         objectMetadata.setContentType(file.getContentType());
         String key = String.format("%s/%d%s", folder, System.currentTimeMillis(), file.getOriginalFilename());
         try {
@@ -38,18 +38,36 @@ public class S3Service {
                     objectMetadata);
             s3client.putObject(putObjectRequest);
         } catch (Exception e) {
-            log.error("Error uploading file to storage", e);
-            throw new RuntimeException("Error uploading file to storage", e);
+            log.error("Ошибка при загрузке файла в хранилище", e);
+            throw new IntegrationException("Ошибка при загрузке файла в хранилище ", e);
         }
         return key;
     }
 
     public void deleteFile(String key) {
-        s3client.deleteObject(new DeleteObjectRequest(bucketName, key));
+        try {
+            if (s3client.doesObjectExist(bucketName, key)) {
+                s3client.deleteObject(new DeleteObjectRequest(bucketName, key));
+                log.info("Файл с ключом {} успешно удалён", key);
+            } else {
+                log.warn("Файл с ключом {} не найден в хранилище", key);
+            }
+        } catch (AmazonServiceException e) {
+            log.error("Ошибка при удалении файла из хранилища: {}", e.getMessage(), e);
+            throw new IntegrationException("Ошибка при удалении файла из хранилища", e);
+        }
     }
 
     public InputStream downloadFile(String key) {
-        S3Object s3object = s3client.getObject(bucketName, key);
-        return s3object.getObjectContent();
+        try {
+            if (!s3client.doesObjectExist(bucketName, key)) {
+                throw new IntegrationException("Файл с ключом " + key + " не найден в хранилище");
+            }
+            S3Object s3object = s3client.getObject(bucketName, key);
+            return s3object.getObjectContent();
+        } catch (AmazonServiceException e) {
+            log.error("Ошибка при загрузке файла из хранилища: {}", e.getMessage(), e);
+            throw new IntegrationException("Ошибка при загрузке файла из хранилища", e);
+        }
     }
 }
